+++ CHANGES.md
@@ -2,6 +2,7 @@
 ======
 * Added support for the PDE editor.
 * Added support for the ColdFusion editor (CFEclipse).
+* Fixed source hover background color in C/C++ editor.
 
 0.10.2 (2011-04-13)
 ===================